Always allocate a NUL-terminated ASCII string for err
There's a few functions that expect a char* (string) but are _not_
guaranteed to initialize it. One should either provide an array of
characters with an explicit length, or a NUL-terminated string. I'm
chosing the latter to strike a balance between memory conservancy
and safety, via `withCAString` (since strings seem to be ASCII
coming from SWE.) We could also use `withCAStringLen`, which seems
to be equivalent to the `allocaArray` approach being replaced here.

See #9 and:

https://www.haskell.org/onlinereport/haskell2010/haskellch27.html