FINALLY FIX THE SEGFAULT
This time, it was due to yet another tricky `static` variable: some
thread set it, implying that the file (e.g. `244`) was open, but the
current thread's file pointer (which was already in thread-local state,)
is NULL, thus throwing it in a situation where it goes and merrily tries
to simply `fseek` in the "open file," only to run into a memory
corruption. I'm not sure how it never broke in macOS, was only able
to reproduce reliably in a Dockerized linux env.